feat(admin): Server Component split for signup/reset-password/mfa/rotate-password/setup (#905)
Applies the Server Component shell + Client Form pattern from PR #901
to the remaining 5 auth pages, fixing the Next.js client-bundle
env-inlining bug across the full Fleet customer auth surface.

Each page.tsx is now a Server Component (no "use client") that reads
env vars at request time and passes resolved values down via props:
- signup/page.tsx → reads NEXT_PUBLIC_API_URL → SignupForm.tsx
- reset-password/page.tsx → no env reads needed → ResetPasswordForm.tsx
- mfa/page.tsx → no env reads needed → MFAForm.tsx
- rotate-password/page.tsx → no env reads needed → RotatePasswordForm.tsx
- setup/page.tsx → reads NEXT_PUBLIC_SITE_NAME → SetupForm.tsx

Additional fixes applied inline (no-risk-too-small):
- rotate-password: replace raw <label> with FormLabel (design system)
- mfa: replace /\D/g regex with filterDigits() (no-regex rule)
- All 5 form components migrated from Card/CardHeader/CardContent to
  Heading + div layout, matching the LoginForm.tsx canonical shape

Closes the PR #901 follow-on TODO.
